+++ nuxeo/nuxeo.py
@@ -4,15 +4,19 @@ from __future__ import unicode_literals
 
 import base64
 import json
+import logging
 import socket
 import tempfile
-import urllib2
 from collections import Sequence
 from urllib import urlencode
+
+import requests
+from requests import HTTPError
 
 from .batchupload import BatchUpload
 from .blob import Blob
 from .directory import Directory
+from .exceptions import Unauthorized
 from .groups import Groups
 from .operation import Operation
 from .repository import Repository
@@ -21,7 +25,11 @@ from .workflow import Workflows
 
 __all__ = ('Nuxeo',)
 
-PARAM_TYPES = {
+logger = logging.getLogger(__name__)
+
+CHUNK_SIZE = 8192  # Chunk size to download files
+
+PARAM_TYPES = {  # Types allowed for operations parameters
     'blob': (unicode, Blob),
     'boolean': (bool,),
     'date': (unicode,),
@@ -58,10 +66,6 @@ def json_helper(o):
     raise TypeError(repr(o) + 'is not JSON serializable (no to_json() found)')
 
 
-class InvalidBatchException(Exception):
-    pass
-
-
 class Nuxeo(object):
     """
     Client for the Nuxeo REST API.
@@ -71,8 +75,6 @@ class Nuxeo(object):
 
     blob_timeout is long (or infinite) timeout dedicated to long HTTP
     requests involving a blob transfer.
-
-    TODO: Switch to Requests to handle proxy
 
     :param base_url: Nuxeo server URL
     :param auth: Authentication parameter {'user': 'Administrator',
@@ -103,7 +105,9 @@ class Nuxeo(object):
         check_suspended=None,
         api_path='api/v1/',
     ):
-        self._headers = {}
+        self._session = requests.session()
+        self._session.proxies = proxies
+        self._session.stream = True
 
         # Function to check during long-running processing like upload /
         # download if the synchronization thread needs to be suspended
@@ -116,8 +120,7 @@ class Nuxeo(object):
         self.blob_timeout = (60 if blob_timeout is None or blob_timeout < 0
                              else blob_timeout)
 
-        self.upload_tmp_dir = (upload_tmp_dir if upload_tmp_dir is not None
-                               else tempfile.gettempdir())
+        self.upload_tmp_dir = upload_tmp_dir or tempfile.gettempdir()
 
         if not base_url.endswith('/'):
             base_url += '/'
@@ -127,15 +130,13 @@ class Nuxeo(object):
 
         self._repository = repository
         self.user_id = None
-        self._auth = ['', '']
-        self._update_auth(auth=auth)
-        self.cookie_jar = cookie_jar
-        cookie_processor = urllib2.HTTPCookieProcessor(
-            cookiejar=cookie_jar)
 
-        # Build URL openers
-        self.streaming_opener = urllib2.build_opener(cookie_processor)
-        self.opener = self.streaming_opener
+        self._auth = {}
+        self._update_auth(auth=auth)
+        self._headers = {'Cache-Control': 'no-cache'}
+
+        if cookie_jar:
+            self._session.cookies = cookie_jar
 
         self.automation_url = base_url + 'site/automation/'
         self.batch_upload_url = 'batch/upload'
@@ -152,8 +153,46 @@ class Nuxeo(object):
         """
         return BatchUpload(self)
 
-    def debug(self, *args, **kwargs):
-        pass
+    def check_params(self, command, params):
+        # type: (unicode, Dict[unicode, Tuple[type, ...]]) -> None
+        """
+        Check given paramaters of the `command` operation.  It will also
+        check for types whenever possible.
+
+        :raises ValueError: When the `command` is not valid.
+        :raises ValueError: On unexpected parameter.
+        :raises ValueError: On missing required parameter.
+        :raises TypeError: When a parameter has not the required type.
+        """
+
+        operation = self.operations.get(command)
+        if not operation:
+            raise ValueError('{!r} is not a registered operation'.format(command))
+
+        parameters = {param['name']: param for param in operation['params']}
+
+        for name, value in params.iteritems():
+            # Check for unexpected parameters.  We use `dict.pop()` to
+            # get and delete the parameter from the dict.
+            try:
+                type_needed = parameters.pop(name)['type']
+            except KeyError:
+                err = 'unexpected parameter {!r} for operation {!r}'
+                raise ValueError(err.format(name, command))
+
+            # Check types
+            types_accepted = PARAM_TYPES.get(type_needed, tuple())
+            if not isinstance(value, types_accepted):
+                err = 'parameter {!r} should be of type {!r} (current {!r})'
+                raise TypeError(err.format(
+                    name, types_accepted, type(name).__name__))
+
+        # Check for required parameters.  As of now, `parameters` may contain
+        # unclaimed parameters and we just need to check for required ones.
+        for name, parameter in parameters.iteritems():
+            if parameter['required']:
+                err = 'missing required parameter {!r} for operation {!r}'
+                raise ValueError(err.format(name, command))
 
     def directory(self, name):
         """
@@ -167,20 +206,11 @@ class Nuxeo(object):
         the $NUXEO_URL/api/v1/drive/configuration endpoint.
         """
 
-        url = self.rest_url + 'drive/configuration'
-        headers = self._get_common_headers()
-        self.trace('Fetching the Drive configuration at %r with headers=%r',
-                   url, headers)
-        req = Request(url, headers=headers)
         try:
-            ret = self.opener.open(req, timeout=self.timeout)
-            return json.loads(ret.read())
-        except (urllib2.URLError, ValueError):
-            pass
+            return self.send(self.rest_url + 'drive/configuration').json()
+        except (HTTPError, ValueError, TypeError):
+            logger.warning('Drive JSON configuration not found.')
         return {}
-
-    def error(self, *args, **kwargs):
-        pass
 
     def execute(
         self,
@@ -203,7 +233,8 @@ class Nuxeo(object):
         :param timeout: Operation timeout
         :param check_params: Verify that the params are valid on the
                              client side
-        :param void_op: If operation is a void operation
+        :param void_op: If True, the response contains no data,
+                        just the status
         :param extra_headers: Headers to add to the request
         :param file_out: Output result inside this file
         :param params: Any additional param to add to the request
@@ -222,13 +253,12 @@ class Nuxeo(object):
             'X-NXproperties': '*',
             # Keep compatibility with old header name
             'X-NXDocumentProperties': '*',
+            'X-NXRepository': self._repository,
         }
         if void_op:
-            headers.update({'X-NXVoidOperation': 'true'})
-        headers.update({'X-NXRepository': self._repository})
+            headers['X-NXVoidOperation'] = 'true'
         if extra_headers:
             headers.update(extra_headers)
-        headers.update(self._get_common_headers())
 
         json_struct = {'params': {}}
         for k, v in params.iteritems():
@@ -236,9 +266,7 @@ class Nuxeo(object):
                 continue
             if k == 'properties':
                 if isinstance(v, dict):
-                    s = ''
-                    for propname, propvalue in v.iteritems():
-                        s += '%s=%s\n' % (propname, propvalue)
+                    s = '\n'.join(['{}={}'.format(name, value) for name, value in v.iteritems()])
                 else:
                     s = v
                 json_struct['params'][k] = s.strip()
@@ -252,43 +280,29 @@ class Nuxeo(object):
                 json_struct['input'] = op_input
         data = json.dumps(json_struct, default=json_helper)
 
-        req = urllib2.Request(url, data, headers)
-        timeout = self.timeout if timeout == -1 else timeout
+        resp = self.send(url, method='POST', extra_headers=headers,
+                         data=data, timeout=timeout)
 
-        try:
-            resp = self.opener.open(req, timeout=timeout)
-        except Exception as e:
-            log_details = self._log_details(e)
-            if isinstance(log_details, tuple):
-                _, _, _, error = log_details
-                if error and error.startswith("Unable to find batch"):
-                    raise InvalidBatchException()
-            raise e
-
-        action = self._get_action()
+        action = self.get_action()
         if action and action.progress is None:
             action.progress = 0
 
-        if file_out is not None:
+        if file_out:
             locker = self.unlock_path(file_out)
             try:
                 with open(file_out, 'wb') as f:
-                    while 'downloading':
-                        # Check if synchronization thread was suspended
-                        if self.check_suspended is not None:
-                            self.check_suspended('File download: %s'
-                                                 % file_out)
-                        buffer_ = resp.read(self.get_download_buffer())
-                        if buffer_ == '':
-                            break
+                    for chunk in resp.iter_content(chunk_size=CHUNK_SIZE):
                         if action:
-                            action.progress += self.get_download_buffer()
-                        f.write(buffer_)
-                return None, file_out
+                            action.progress += CHUNK_SIZE
+                        f.write(chunk)
+                return file_out
             finally:
                 self.lock_path(file_out, locker)
         else:
-            return self._read_response(resp, url)
+            try:
+                return resp.json()
+            except ValueError:
+                return resp.content
 
     def groups(self):
         """
@@ -307,16 +321,30 @@ class Nuxeo(object):
 
     def headers(self, extras=None):
         """
-        Return the headers that will be sent to the server
-        You can set additional headers with extras argument.
+        Headers to include in every HTTP requests.
+
+        Includes the authentication heads (token based or basic auth if no
+        token).
+
+        Also include an application name header to make it possible for the
+        server to compute access statistics for various client types (e.g.
+        browser vs devices).
+
+        You can set additional headers with the extras argument.
 
         :param extras: -- a dictionary or object of additional headers to set
         """
-        headers = self._get_common_headers()
+        headers = self._headers.copy()
+
         if extras:
-            self._headers.update(extras)
-        headers.update(self._headers)
+            headers.update(extras)
+        if self._auth:
+            headers.update(self._auth)
+
         return headers
+
+    def lock_path(self, file_out, locker):
+        pass
 
     def login(self):
         """
@@ -385,34 +413,25 @@ class Nuxeo(object):
         """
 
         url = self.rest_url + relative_url
-        if adapter is not None:
+        if adapter:
             url += '/@' + adapter
 
-        if query_params:
-            url += '?' + urlencode(query_params)
-
-        if body is not None and not isinstance(body, str) and not raw_body:
+        if body and not isinstance(body, bytes) and not raw_body:
             body = json.dumps(body, default=json_helper)
 
         headers = {
             'Content-Type': content_type,
             'Accept': 'application/json+nxentity, */*',
         }
-        headers.update(self._get_common_headers())
-        if extra_headers is not None:
+        if extra_headers:
             headers.update(extra_headers)
-        cookies = self._get_cookies()
-        self.trace('Calling REST API %s with headers %r and cookies %r',
-                   url, headers, cookies)
-        req = Request(url, headers=headers, method=method, data=body)
-        timeout = self.timeout if timeout == -1 else timeout
-        try:
-            resp = self.opener.open(req, timeout=timeout)
-        except Exception as e:
-            self._log_details(e)
-            raise e
 
-        return self._read_response(resp, url)
+        resp = self.send(url, method=method, params=query_params,
+                         extra_headers=headers, data=body, timeout=timeout)
+        try:
+            return resp.json()
+        except ValueError:
+            return resp.content
 
     def request_authentication_token(
         self,
@@ -430,8 +449,6 @@ class Nuxeo(object):
         for future login
         """
 
-        err = ('Failed to connect to Nuxeo server {} with user {}'
-               ' to acquire a token').format(self.base_url, self.user_id)
         parameters = {
             'deviceId': device_id,
             'applicationName': application_name,
@@ -443,54 +460,72 @@ class Nuxeo(object):
         url = self.base_url + 'authentication/token?'
         url += urlencode(parameters)
 
-        headers = self._get_common_headers()
-        cookies = self._get_cookies()
-        self.trace('Calling %s with headers %r and cookies %r',
-                   url, headers, cookies)
-        req = urllib2.Request(url, headers=headers)
-        try:
-            token = self.opener.open(req, timeout=self.timeout).read()
-        except urllib2.HTTPError as e:
-            if e.code == 401 or e.code == 403:
-                raise Unauthorized(self.base_url, self.user_id, e.code)
-            elif e.code == 404:
-                # Token based auth is not supported by this server
-                return None
-            else:
-                e.msg = err + ': HTTP error %d' % e.code
-                raise e
-        except Exception as e:
-            if hasattr(e, 'msg'):
-                e.msg = err + ': ' + e.msg
-            raise
-        cookies = self._get_cookies()
-        self.trace('Got token %r with cookies %r', token, cookies)
+        token = self.send(url).text
+
         # Use the (potentially re-newed) token from now on
         if not revoke:
             self._update_auth(token=token)
         return token
+
+    def send(self, url, method='GET', data=None, params=None, extra_headers=None, timeout=None):
+        """
+        Perform a request to the server.
+
+        This method acts as a wrapper for the request to handle the errors
+        and the logging. All other methods use this one to send their requests.
+
+        :param url: URL for the HTTP request
+        :param method: Method of the HTTP request
+        :param data: Body of the request (text or bytes)
+        :param params: Parameters to encode in the URL
+        :param extra_headers: Headers to add to the common ones
+        :param timeout: Timeout for the request
+        :return: The HTTP request response (Response object)
+        """
+        if method not in ('GET', 'HEAD', 'POST', 'PUT',
+                          'DELETE', 'CONNECT', 'OPTIONS', 'TRACE'):
+            raise ValueError('method parameter is not a valid HTTP method.')
+
+        timeout = self.timeout if not timeout or timeout == -1 else timeout
+        headers = self.headers(extra_headers)
+
+        logger.debug('Calling {!r} with headers {!r} and cookies {!r}'.format(
+            url, headers, self._get_cookies()))
+
+        try:
+            resp = self._session.request(url=url, method=method, headers=headers,
+                                         params=params, data=data, timeout=timeout)
+            resp.raise_for_status()
+        except Exception as e:
+            if isinstance(e, HTTPError) and e.response.status_code in (401, 403):
+                e = Unauthorized(self.user_id, e)
+            self._log_details(e)
+            raise e
+
+        if int(resp.headers.get('content-length', CHUNK_SIZE + 1)) <= CHUNK_SIZE:
+            content = resp.content
+        else:
+            content = '<Too much data to display>'
+
+        logger.debug('Response from {!r}: {!r} with cookies {!r}'.format(
+            url, content, self._get_cookies()))
+
+        return resp
 
     def server_reachable(self):
         """
         Simple call to the server status page to check if it is reachable.
         """
 
-        url = self.base_url + 'runningstatus'
-        headers = self._get_common_headers()
-        self.trace('Checking server availability at %r with headers=%r',
-                   url, headers)
-        req = urllib2.Request(url, headers=headers)
         try:
-            ret = self.opener.open(req, timeout=self.timeout)
-        except urllib2.URLError:
+            resp = self.send(self.base_url + 'runningstatus')
+            return resp.ok
+        except HTTPError:
             pass
-        else:
-            if ret.code == 200:
-                return True
         return False
 
-    def trace(self, *args, **kwargs):
-        pass
+    def unlock_path(self, file_out):
+        return None
 
     def users(self):
         """
@@ -504,178 +539,54 @@ class Nuxeo(object):
         """
         return Workflows(self)
 
-    def check_params(self, command, params):
-        # type: (unicode, Dict[unicode, Tuple[type, ...]]) -> None
-        """
-        Check given paramaters of the `command` operation.  It will also
-        check for types whenever possible.
-
-        :raises ValueError: When the `command` is not valid.
-        :raises ValueError: On unexpected parameter.
-        :raises ValueError: On missing required parameter.
-        :raises TypeError: When a parameter has not the required type.
-        """
-
-        operation = self.operations.get(command)
-        if not operation:
-            raise ValueError('%r is not a registered operation' % command)
-
-        parameters = {param['name']: param for param in operation['params']}
-
-        for name, value in params.iteritems():
-            # Check for unexpected paramaters.  We use `dict.pop()` to
-            # get and delete the parameter from the dict.
-            try:
-                type_needed = parameters.pop(name)['type']
-            except KeyError:
-                err = 'unexpected parameter %r for operation %r'
-                raise ValueError(err, name, command)
-
-            # Check types
-            types_accepted = PARAM_TYPES.get(type_needed, tuple())
-            if not isinstance(value, types_accepted):
-                err = 'parameter {!r} should be of type {!r} (current %s)'
-                raise TypeError(err.format(
-                    name, types_accepted, type(name).__name__))
-
-        # Check for required parameters.  As of now, `parameters` may contain
-        # unclaimed parameters and we just need to check for required ones.
-        for name, parameter in parameters.iteritems():
-            if parameter['required']:
-                err = 'missing required parameter {!r} for operation {!r}'
-                raise ValueError(err.format(name, command))
-
-    def _create_action(self, type, path, name):
+    def create_action(self, type, path, name):
         return {}
 
-    def _end_action(self):
+    def end_action(self):
         pass
 
     def _fetch_api(self):
         """ Used to populate :attr:`operations`, do not call directly. """
 
-        err = 'Failed to connect to Nuxeo server {}'.format(self.base_url)
-        url = self.automation_url
-        headers = self._get_common_headers()
-        cookies = self._get_cookies()
-        self.trace('Calling %s with headers %r and cookies %r',
-                   url, headers, cookies)
-        req = urllib2.Request(url, headers=headers)
-        try:
-            response = json.loads(
-                self.opener.open(req, timeout=self.timeout).read())
-        except urllib2.HTTPError as e:
-            if e.code in (401, 403):
-                raise Unauthorized(self.base_url, self.user_id, e.code)
-
-            msg = err + '\nHTTP error %d' % e.code
-            if hasattr(e, 'msg'):
-                msg = msg + ': ' + e.msg
-            e.msg = msg
-            raise e
-        except urllib2.URLError as e:
-            msg = err
-            if hasattr(e, 'message') and e.message:
-                e_msg = force_decode(': ' + e.message)
-                if e_msg is not None:
-                    msg += e_msg
-            elif hasattr(e, 'reason') and e.reason:
-                if (hasattr(e.reason, 'message')
-                        and e.reason.message):
-                    e_msg = force_decode(': ' + e.reason.message)
-                    if e_msg is not None:
-                        msg += e_msg
-                elif (hasattr(e.reason, 'strerror')
-                        and e.reason.strerror):
-                    e_msg = force_decode(': ' + e.reason.strerror)
-                    if e_msg is not None:
-                        msg += e_msg
-            msg += ('\nPlease check your Internet connection,'
-                    + ' make sure the Nuxeo server URL is valid'
-                    + ' and check your proxy settings.')
-            e.msg = msg
-            raise e
-        except Exception as e:
-            msg = err
-            if hasattr(e, 'msg'):
-                msg += ': ' + e.msg
-            e.msg = msg
-            raise e
+        resp = self.send(self.automation_url)
+        resp = resp.json()
 
         operations = {}
-        for operation in response['operations']:
+        for operation in resp['operations']:
             operations[operation['id']] = operation
             for alias in operation.get('aliases', []):
                 operations[alias] = operation
         return operations
 
-    def _get_action(self):
+    def get_action(self):
         return None
-
-    def _get_common_headers(self):
-        """
-        Headers to include in every HTTP requests.
-
-        Includes the authentication heads (token based or basic auth if no
-        token).
-
-        Also include an application name header to make it possible for the
-        server to compute access statistics for various client types (e.g.
-        browser vs devices).
-        """
-
-        headers = {
-            'Cache-Control': 'no-cache',
-        }
-        if self._auth is not None:
-            headers.update([self._auth])
-        headers.update(self._headers)
-        return headers
 
     def _get_cookies(self):
-        return list(self.cookie_jar) if self.cookie_jar is not None else []
+        return list(self._session.cookies) or []
 
     def _log_details(self, e):
-        if hasattr(e, 'fp'):
-            detail = e.fp.read().decode('utf-8')
+        if isinstance(e, HTTPError):
+            logger.exception(u'Remote exception: {}'.format(
+                e.message.decode('utf-8')))
             try:
-                exc = json.loads(detail)
+                exc = e.response.json()
                 message = exc.get('message')
-                stack = exc.get('stack')
+                stack = exc.get('stacktrace')
                 error = exc.get('error')
                 if message:
-                    self.debug('Remote exception message: %s', message)
+                    logger.error('Remote exception message: {!s}'.format(message))
                 if stack:
-                    self.debug('Remote exception stack: %r',
-                               exc['stack'], exc_info=True)
+                    logger.error('Remote exception stack: {!s}'.format(stack))
                 else:
-                    self.debug('Remote exception details: %r', detail)
+                    logger.error('Remote exception details: {!s}'.format(exc))
                 return exc.get('status'), exc.get('code'), message, error
             except ValueError:
-                # Error message should always be a JSON message,
-                # but sometimes it's not
-                if '<html>' in detail:
-                    message = e
-                else:
-                    message = detail
-                self.error(message)
-                if isinstance(e, urllib2.HTTPError):
-                    return e.code, None, message, None
-        return None
-
-    def _read_response(self, response, url):
-        info = response.info()
-        s = response.read()
-        content_type = info.get('content-type', '')
-        cookies = self._get_cookies()
-        if content_type.startswith('application/json'):
-            self.trace('Response for %r with cookies %r: %r',
-                       url, cookies, s)
-            return json.loads(s) if s else None
+                # Error messages from the server should always be JSON-formatted,
+                # but sometimes they're not
+                logger.error('Response is not JSON')
         else:
-            self.trace('Response for %r with cookies %r has content-type %r',
-                       url, cookies, content_type)
-            return s
+            # The error was not sent from the server
+            logger.exception('Local exception')
 
     def _update_auth(self, auth=None, password=None, token=None):
         """
@@ -685,46 +596,17 @@ class Nuxeo(object):
         to string.
         """
 
-        if auth is not None:
-            if 'username' in auth:
-                self.user_id = auth['username']
-            if 'token' in auth:
-                token = auth['token']
-            if 'password' in auth:
-                password = auth['password']
-
-        if self.user_id and isinstance(self.user_id, unicode):
-            self.user_id = unicode(self.user_id).encode('utf-8')
+        if auth:
+            self.user_id = auth.get('username', None)
+            token = auth.get('token', token)
+            password = auth.get('password', password)
 
         # Select the most appropriate auth headers based on credentials
-        if token is not None:
-            self._auth = ('X-Authentication-Token', token)
-        elif password is not None:
-            basic_auth = 'Basic %s' % base64.b64encode(
-                    self.user_id + ":" + password).strip()
-            self._auth = 'Authorization', basic_auth
+        if token:
+            self._auth = {'X-Authentication-Token': token}
+        elif password:
+            self._auth = {'Authorization': 'Basic {}'.format(
+                base64.b64encode('{}:{}'.format(
+                    self.user_id, password).encode('utf-8')).strip())}
         else:
             raise ValueError('Either password or token must be provided')
-
-
-class Request(urllib2.Request):
-    """ Need to override urllib2 request to add the HTTP method. """
-
-    def __init__(self, *args, **kwargs):
-        self._method = kwargs.pop('method', 'GET')
-        urllib2.Request.__init__(self, *args, **kwargs)
-
-    def get_method(self):
-        return self._method
-
-
-class Unauthorized(Exception):
-
-    def __init__(self, server_url, user_id, code=403):
-        self.server_url = server_url
-        self.user_id = user_id
-        self.code = code
-
-    def __str__(self):
-        return ('%r is not authorized to access %r with '
-                ' the provided credentials' % (self.user_id, self.server_url))