+++ tests/test_nuxeo.py
@@ -1,11 +1,16 @@
 # coding: utf-8
 from __future__ import unicode_literals
 
-from io import BufferedIOBase
-from urllib2 import HTTPError, URLError
+import json
+import os
 
 import pytest
+import requests
+from requests import HTTPError
+
 from nuxeo.blob import Blob
+from nuxeo.exceptions import Unauthorized
+from nuxeo.nuxeo import Nuxeo
 
 
 @pytest.mark.parametrize('method, params, is_valid', [
@@ -22,8 +27,8 @@ from nuxeo.blob import Blob
     ('Document.Fetch', {'value': True}, False),
     ('Document.MultiPublish', {'target': ['test1.txt', 'test2.txt']}, True),  # 'target' type == list
     ('Document.MultiPublish', {'target': 0}, False),
-    ('Picture.Resize', {'maxHeight': 100, 'maxWidth': 100}, True),  # 'maxHeight', 'maxWidth' type == int
-    ('Picture.Resize', {'maxHeight': 'test', 'maxWidth': 'test'}, False),
+    ('Audit.Query', {'query': 'test', 'pageNo': 100}, True),  # 'pageNo' type == int
+    ('Audit.Query', {'query': 'test', 'pageNo': 'test'}, False),
     ('Document.Query', {'query': 'test', 'pageSize': 10}, True),  # 'pageSize' type == int
     ('Document.Query', {'query': 'test', 'pageSize': 'test'}, False),
     ('PDF.ExtractPages', {'startPage': 1, 'endPage': 2}, True),  # 'startPage', 'endPage' type == long
@@ -61,10 +66,15 @@ def test_check_params_unknown_operation(server):
 def test_drive_config(monkeypatch, server):
 
     def mock_server_error(*args, **kwargs):
-        raise URLError('Mock error')
+        def mock_HTTP():
+            return HTTPError('Mock error')
+        mock_response = requests.Response()
+        mock_response.status_code = 404
+        mock_response.raise_for_status = mock_HTTP
+        return mock_response
 
     def mock_invalid_response(*args, **kwargs):
-        return BufferedIOBase()
+        return requests.Response()
 
     config = server.drive_config()
     assert isinstance(config, dict)
@@ -77,10 +87,10 @@ def test_drive_config(monkeypatch, server):
         assert 'update_check_delay' in config
         assert 'ui' in config
 
-    monkeypatch.setattr(server.opener, 'open', mock_server_error)
+    monkeypatch.setattr(requests.Session, 'request', mock_server_error)
     assert not server.drive_config()
     monkeypatch.undo()
-    monkeypatch.setattr(server.opener, 'open', mock_invalid_response)
+    monkeypatch.setattr(requests.Session, 'request', mock_invalid_response)
     assert not server.drive_config()
     monkeypatch.undo()
 
@@ -92,9 +102,21 @@ def test_encoding_404_error(server):
     try:
         with pytest.raises(HTTPError) as e:
             server.repository().fetch('/')
-        assert e.value.code == 404
+        assert e.value.response.status_code == 404
     finally:
         server.rest_url = rest_url
+
+
+def test_file_out(server):
+    operation = server.operation('Document.GetChild')
+    operation.params({'name': 'workspaces'})
+    operation.input('/default-domain')
+    file_out = operation.execute(file_out='test')
+    with open(file_out, 'rb') as f:
+        file_content = json.loads(f.read())
+        resp_content = operation.execute()
+        assert file_content == resp_content
+    os.remove(file_out)
 
 
 def test_get_operations(server):
@@ -127,3 +149,25 @@ def test_server_reachable(server):
         assert not server.server_reachable()
     finally:
         server.base_url = base_url
+
+
+def test_unauthorized(server):
+    credentials = {
+        'username': 'ミカエル',
+        'password': 'test',
+    }
+    user = server.users().create(credentials)
+
+    try:
+        new_server = Nuxeo(
+            base_url=os.environ.get('NXDRIVE_TEST_NUXEO_URL',
+                                    'http://localhost:8080/nuxeo'),
+            auth=credentials)
+
+        with pytest.raises(Unauthorized):
+            new_server.users().create({
+                'username': 'another_one',
+                'password': 'test'
+            })
+    finally:
+        user.delete()